% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{os_alert}
\alias{os_alert}
\title{Show alert dialog}
\usage{
os_alert(msg)
}
\arguments{
\item{msg}{message}
}
\description{
\code{os_alert} evoke OS native alert dialog
}
\examples{
# os_alert("Hello!")
}
