% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{os_dialog}
\alias{os_dialog}
\title{Show OS native dialog}
\usage{
os_dialog(msg_obj)
}
\arguments{
\item{msg}{message}
}
\description{
\code{os_dialog} evoke OS native dialog
}
\examples{
# os_dialog(list(message = "Hello",continue = "OK",abort = "Cancel"))
}
